Import for custom leagues Fixes #577
